[None][fix] Revert unnecessary int32->int64 change for kv_cache_pool_mapping
V2 KVCacheManager (C++) creates kv_cache_pool_mapping with SizeType32
(int32). The int64 change was unnecessary and introduced a dtype
mismatch. Revert both Python-side (resource_manager.py) and C++-side
(attentionOp.cpp/h) back to int32.

+++ tensorrt_llm/_torch/pyexecutor/resource_manager.py
@@ -1908,7 +1908,7 @@ class KVCacheManagerV2(BaseResourceManager):
             kv_cache_pool_mapping_list.append([layer_group_id, offset])
 
         kv_cache_pool_mapping = torch.tensor(kv_cache_pool_mapping_list,
-                                             dtype=torch.int64,
+                                             dtype=torch.int32,
                                              device="cpu",
                                              pin_memory=prefer_pinned())
         return kv_cache_pool_pointers, kv_cache_pool_mapping
